Add files via upload
This version of the stomploader.sh script loads the pru_adc.out and pru_osc.out firmwares on the PRU1 and PRU0, respectively, before launching the Qt application. This ensures that the scope application is working right out-of-the-box. At the previous version, the firmwares would be loaded manually by the load.sh script within a SSH shell.