Add files via upload
Working time_div and volt_div with generic definitions of axis limits, magnification and frame rate. Also, the encoder debouncing was introduced and it worked properly. The next step is to fill the entire frame when the time_div falls down. This will be done by changing the IEP timer compare register, via the Qt application.