Add files via upload
When the time_div decreases a lot, you got the sensation of a infinity of continuous points on screen. The frame has a maximum number of points that are plotted at each update, depending on the panel width in pixels and on the desired number of points per pixel. The frame is only updated when all the points are treated and formatted in the shape of a QVector. Also, the capture and treatment of data is kept in a separated thread from its actual plot process on screen. The next challenge is to come up with some way of plotting data as the QVector is filled (nobody knows how to do it yet, without subclassing QCharts's classes or rebuilding Qt itself), as well as implementing the trigger system.