Add files via upload
First working trigger system. There are still some bugs, concerning the frontiers of the frame, when the offset is negative. The offset is the distance between the trigger point and the 'x' coordinate of the center of the screen. The fix is to change the way we deal with the acquisition buffer. The frame should lie upon the buffer and the buffer must exceed the frame in both directions, in such a way that the frontier bugs would disappear.