Add files via upload
The sample rate was kept always at POINTS_PER_FRAME/time_div, respecting the MAX_SAMPLE_RATE constraint, which is currently 300 kHz per channel. The PACK_SIZE, ENCODER_RATE and FRAME_RATE were tuned in a way that the encoder's response is very fast and the screen still updates very rapidly, even on slow windows. Also, the MAG_X_MAX has been modified to 0,1 s, which is enough time to display 2 periods of a 20 Hz wave. 

The next step is to discover why the Qt application seems to misbehave on measuring slow signals (< 1 kHz). Maybe it is a hardware problem, risen by the poor use of 555 timer, that won't be verified by plugging in a true signal generator. The second next step is to dive into the shader subject.