Add files via upload
The data compress routine was eliminated. Now, the nice_update_requests are meant for windows that were triggered and raw_update_requests are meant for windows that weren't triggered. A lot of minor bugs were fixed, regarding the indexing of the shifting operations demanded by the trigger system. 

The sample's buffer now has the largest size possible: 1920 samples for each channel, which means 15 kB of data, which is basically the PRU's limit. This enhances the Qt application's performance on low frequency signals, because the buffer is now capable of representing twice the time it could before. Nevertheless, there is still one performance gap, which is basically impossible to overcome: no matter how fast the PRU's firmware updates the buffer nor how deep the buffer is if the Qt application isn't capable of returning to the data_tick_timeout thread and actually read the buffer. We reached our performance frontier, for now.