Add files via upload
Each time the QTLAZY flag is held, the PRU's firmware acquires 1920 samples, which are divided into ARMx_BUFFERA and ARMx_BUFFERB for both channels. The core of the whole scope software is the acquisition made by PRU1. In order to make it faster, STEPx_BUFFER occupies the whole PRU0's RAM. The ARM1_BUFFERA and ARM1_BUFFERB buffers occupy the PRU1's RAM and the ARM2_BUFFERA and ARM2_BUFFERB occupy the shared RAM.