Add files via upload
Leaner code. Global variables were kept to a minimum. Global objects were encapsulated as MyMainWindow's members. Every object construction was made by using new[], which ensures that any destruction order can be taken. Dynamic allocations are now only made by using new[] and the memory is freed as soon as possible by the delete[] command, except for the encoders' system, which is heavily "C" based. We look forward to review that in a medium-term future.